[level 0] Title: 등수 매기기, Time: 0.52 ms, Memory: 74.2 MB -BaekjoonHub

+++ 프로그래머스/0/120882. 등수 매기기/등수 매기기.java
@@ -1,25 +1,24 @@
 import java.util.*;
-
 class Solution {
     public int[] solution(int[][] score) {
-        double[] evtemp = new double[score.length];
-        double[] answer = new double[score.length];
-
+        int[] evtemp = new int[score.length];
+        int[] answer = new int[score.length];
+        
         for (int i = 0; i < score.length; i++) {
-            evtemp[i] = (score[i][0] + score[i][1]) / 2.0;
+            evtemp[i] = (score[i][0] + score[i][1]) / 2;
             answer[i] = evtemp[i];
         }
-
+        
         Arrays.sort(evtemp);
         for (int i = 0; i < evtemp.length / 2; i++) {
-            double temp = evtemp[i];
+            int temp = evtemp[i];
             evtemp[i] = evtemp[evtemp.length - 1 - i];
             evtemp[evtemp.length - 1 - i] = temp;
         }
-
+        
         int[] rank = new int[score.length];
         int currentRank = 1;
-
+        
         for (int i = 0; i < evtemp.length; i++) {
             for (int j = 0; j < score.length; j++) {
                 if (evtemp[i] == answer[j] && rank[j] == 0) {
@@ -28,7 +27,7 @@ class Solution {
             }
             currentRank++;
         }
-
+        
         return rank;
     }
 }
